api/requirements.txt
fastapi
line-bot-sdk
pydantic-settings
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
psycopg2-binary
uvicorn
requests